refactor(#122): split core/ptm_constants.py into per-plugin and per-algorithm homes
- UNIPROT_* -> skills/uniprot_ptm/shared/constants.py
- PEPTIDEATLAS_* -> skills/peptideatlas/phospho/shared/constants.py
- CPTAC_* -> skills/cptac/shared/constants.py
- ENSEMBL_GTF_*/RELEASE + PTM_*/LMM_*/RESOLUTION_*/EXPRESSION_*/etc.
  -> algorithms/ptm/constants.py (NEW). Correction to the issue table:
  ENSEMBL_GTF_* are consumed only by algorithms/ptm/pipeline.py
  (download_ensembl_gtf), not by the peptideatlas plugin, so they belong
  with the PTM algorithm.
- Removed the dead UNIPROT_* re-exports from algorithms/ptm/__init__.py
  (no external consumer; re-pointing them to skills/ would violate the
  algorithms->skills rule).
- Deleted the now-empty core/ptm_constants.py.

One of several PRs closing #122.
