+++ hvantk/skills/peptideatlas/phospho/shared/constants.py
@@ -1,0 +1,9 @@
+"""Plugin-local constants for the peptideatlas skill.
+
+Moved from hvantk/core/ptm_constants.py per issue #122 (clean-core principle).
+"""
+
+# PeptideAtlas Phospho Build
+PEPTIDEATLAS_PHOSPHO_BASE_URL = "https://peptideatlas.org/builds/human/phospho"
+PEPTIDEATLAS_LATEST_BUILD_DATE = "202512"
+PEPTIDEATLAS_LATEST_BUILD_ID = "606"
